Combat Manager

+++ src/main/java/main/combat/CombatCalculator.java
@@ -10,12 +10,12 @@ public class CombatCalculator {
     public static CombatCalculator getInstance(){
         return INSTANCE;
     }
-    public int calculateDamage(Character atacante, Character defensor){
-        int resultado = (int)(atacante.getFuerza() * atacante.getAgilidad()*0.5 - defensor.getResistencia()*0.2);
+    public double calculateDamage(Character atacante, Character defensor){
+        double resultado = (atacante.getFuerza() * atacante.getAgilidad()*0.5 - defensor.getResistencia()*0.2);
         return Math.max(1, resultado);
     }
-    public int calculateBlock(Character defensor){
-        int resultado = (int)(defensor.getResistencia()*0.5);
+    public double calculateBlock(Character defensor){
+        double resultado = (defensor.getResistencia()*0.5);
         return Math.max(1, resultado);
     }
 }
